src/requirements.txt
pika==0.12.0
pytest-xdist
sqlalchemy
psycopg2
